% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bvarPANELs-package.R
\docType{package}
\name{bvarPANELs-package}
\alias{bvarPANELs-package}
\alias{bvarPANELs}
\title{Forecasting with Bayesian Hierarchical Panel Vector Autoregressions}
\description{
Forecasting a multi-country time series panel data using 
Bayesian Vector Autoregressions with a three-level country-global 
hierarchical prior structure. Copyright: 2024 International Labour Organization.
}
\note{
This package is currently in active development.
}
\author{
Tomasz Woźniak \email{wozniak.tom@pm.me}
}
\keyword{models}
\keyword{package}
\keyword{ts}
