% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate.bvarPANEL.R
\name{estimate.PosteriorBVARPANEL}
\alias{estimate.PosteriorBVARPANEL}
\title{Bayesian estimation of a Bayesian Hierarchical Panel Vector 
Autoregression using Gibbs sampler}
\usage{
\method{estimate}{PosteriorBVARPANEL}(specification, S, thin = 1, show_progress = TRUE)
}
\arguments{
\item{specification}{an object of class PosteriorBVARPANEL generated using 
the \code{estimate.BVARPANEL()} function. This setup facilitates the 
continuation of the MCMC sampling starting from the last draw of the previous 
run.}

\item{S}{a positive integer, the number of posterior draws to be generated}

\item{thin}{a positive integer, specifying the frequency of MCMC output thinning}

\item{show_progress}{a logical value, if \code{TRUE} the estimation progress 
bar is visible}
}
\value{
An object of class PosteriorBVARPANEL containing the Bayesian 
estimation output and containing two elements:

 \code{posterior} a list with a collection of \code{S} draws from the 
 posterior distribution generated via Gibbs sampler.
\code{last_draw} an object of class BVARPANEL with the last draw of the 
current MCMC run as the starting value to be passed to the continuation of 
the MCMC estimation using the \code{estimate()} method.
}
\description{
Estimates the Bayesian Hierarchical Panel VAR using the Gibbs 
sampler proposed by Sanchez-Martinez & Woźniak (2024).
}
\details{
The homoskedastic SVAR model is given by the reduced form equation:
\deqn{Y_c = A_cX_c + E_c}
where \eqn{Y_c} is an \code{T_c x N} matrix of dependent variables for 
country \code{c}, \eqn{X_c} is a \code{T_c x K} matrix of explanatory 
variables, \eqn{E_c} is an \code{T_c x N} matrix of error terms, and 
\eqn{A_c} is an \code{NxK} matrix of country-specific autoregressive slope 
coefficients and parameters on deterministic terms in \eqn{X_c}.
}
\examples{
data(ilo_cubic_panel)                                   # load the data
data(ilo_exogenous_variables)                           # load the exogenous variables
set.seed(123)
# specify the model
specification = specify_bvarPANEL$new(ilo_cubic_panel, exogenous = ilo_exogenous_variables)
burn_in       = estimate(specification, 10)             # run the burn-in
posterior     = estimate(burn_in, 10)                   # estimate the model

}
\seealso{
\code{\link{specify_bvarPANEL}}, \code{\link{specify_posterior_bvarPANEL}}
}
\author{
Tomasz Woźniak \email{wozniak.tom@pm.me}
}
