% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forecast.R
\name{forecast.PosteriorBVARPANEL}
\alias{forecast.PosteriorBVARPANEL}
\title{Forecasting using Hierarchical Pannel Vector Autoregressions}
\usage{
\method{forecast}{PosteriorBVARPANEL}(posterior, horizon = 1, exogenous_forecast = NULL)
}
\arguments{
\item{posterior}{posterior estimation outcome - an object of class 
PosteriorBVARPANEL obtained by running the \code{estimate} function.}

\item{horizon}{a positive integer, specifying the forecasting horizon.}

\item{exogenous_forecast}{not used here ATM; included for compatibility with 
generic \code{forecast}.}
}
\value{
A list of class \code{PanelForecasts} containing the
draws from the predictive density and data. The output list includes element:

\describe{
 \item{forecasts}{an \code{horizonxNxCxS} array with the draws from predictive density}
 \item{forecasts_cpp}{an unspecified object passed for computations in **cpp**}
 \item{Y}{a \code{C}-element list with \code{T_cxN} matrices with the country-specific data}
}
}
\description{
Samples from the joint predictive density of the dependent 
variables for all countries at forecast horizons 
from 1 to \code{horizon} specified as an argument of the function.
}
\examples{
data(ilo_cubic_panel)                                   # load the data
set.seed(123)
specification = specify_bvarPANEL$new(ilo_cubic_panel)  # specify the model
burn_in       = estimate(specification, 10)             # run the burn-in
posterior     = estimate(burn_in, 10)                   # estimate the model
predictive    = forecast(posterior, 2)                  # forecast 2 years ahead

# workflow with the pipe |>
############################################################
set.seed(123)
ilo_cubic_panel |>
  specify_bvarPANEL$new() |>
  estimate(S = 10) |> 
  estimate(S = 20) |> 
  forecast(horizon = 2) -> predictive

}
\author{
Tomasz Woźniak \email{wozniak.tom@pm.me}
}
