% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/specify_bvarpanel.R
\name{specify_posterior_bvarPANEL}
\alias{specify_posterior_bvarPANEL}
\title{R6 Class Representing PosteriorBVARPANEL}
\description{
The class PosteriorBVARPANEL contains posterior output and the specification 
including the last MCMC draw for the Bayesian Panel VAR model. 
Note that due to the thinning of the MCMC output the starting value in element 
\code{last_draw} might not be equal to the last draw provided in 
element \code{posterior}.
}
\examples{
\dontrun{
# This is a function that is used within estimate()
data(us_fiscal_lsuw)
specification  = specify_bsvar$new(us_fiscal_lsuw, p = 4)
set.seed(123)
estimate       = estimate(specification, 50)
class(estimate)
}

## ------------------------------------------------
## Method `specify_posterior_bvarPANEL$get_posterior`
## ------------------------------------------------

\dontrun{
data(us_fiscal_lsuw)
specification  = specify_bsvar$new(us_fiscal_lsuw)
set.seed(123)
estimate       = estimate(specification, 50)
estimate$get_posterior()
}

## ------------------------------------------------
## Method `specify_posterior_bvarPANEL$get_last_draw`
## ------------------------------------------------

\dontrun{
data(us_fiscal_lsuw)

# specify the model and set seed
specification  = specify_bsvar$new(us_fiscal_lsuw, p = 4)
set.seed(123)

# run the burn-in
burn_in        = estimate(specification, 10)

# estimate the model
posterior      = estimate(burn_in, 10)
}
}
\seealso{
\code{\link{specify_bvarPANEL}}
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{last_draw}}{an object of class BVARPANEL with the last draw of the 
current MCMC run as the starting value to be passed to the continuation 
of the MCMC estimation using \code{estimate()}.}

\item{\code{posterior}}{a list containing Bayesian estimation output.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-PosteriorBVARPANEL-new}{\code{specify_posterior_bvarPANEL$new()}}
\item \href{#method-PosteriorBVARPANEL-get_posterior}{\code{specify_posterior_bvarPANEL$get_posterior()}}
\item \href{#method-PosteriorBVARPANEL-get_last_draw}{\code{specify_posterior_bvarPANEL$get_last_draw()}}
\item \href{#method-PosteriorBVARPANEL-clone}{\code{specify_posterior_bvarPANEL$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-PosteriorBVARPANEL-new"></a>}}
\if{latex}{\out{\hypertarget{method-PosteriorBVARPANEL-new}{}}}
\subsection{Method \code{new()}}{
Create a new posterior output PosteriorBVARPANEL.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{specify_posterior_bvarPANEL$new(specification_bvarPANEL, posterior_bvarPANEL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{specification_bvarPANEL}}{an object of class BVARPANEL with the last 
draw of the current MCMC run as the starting value.}

\item{\code{posterior_bvarPANEL}}{a list containing Bayesian estimation output.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A posterior output PosteriorBVARPANEL.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-PosteriorBVARPANEL-get_posterior"></a>}}
\if{latex}{\out{\hypertarget{method-PosteriorBVARPANEL-get_posterior}{}}}
\subsection{Method \code{get_posterior()}}{
Returns a list containing Bayesian estimation output.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{specify_posterior_bvarPANEL$get_posterior()}\if{html}{\out{</div>}}
}

\subsection{Examples}{
\if{html}{\out{<div class="r example copy">}}
\preformatted{\dontrun{
data(us_fiscal_lsuw)
specification  = specify_bsvar$new(us_fiscal_lsuw)
set.seed(123)
estimate       = estimate(specification, 50)
estimate$get_posterior()
}
}
\if{html}{\out{</div>}}

}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-PosteriorBVARPANEL-get_last_draw"></a>}}
\if{latex}{\out{\hypertarget{method-PosteriorBVARPANEL-get_last_draw}{}}}
\subsection{Method \code{get_last_draw()}}{
Returns an object of class BVARPANEL with the last draw of the current 
MCMC run as the starting value to be passed to the continuation of the 
MCMC estimation using \code{estimate()}.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{specify_posterior_bvarPANEL$get_last_draw()}\if{html}{\out{</div>}}
}

\subsection{Examples}{
\if{html}{\out{<div class="r example copy">}}
\preformatted{\dontrun{
data(us_fiscal_lsuw)

# specify the model and set seed
specification  = specify_bsvar$new(us_fiscal_lsuw, p = 4)
set.seed(123)

# run the burn-in
burn_in        = estimate(specification, 10)

# estimate the model
posterior      = estimate(burn_in, 10)
}
}
\if{html}{\out{</div>}}

}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-PosteriorBVARPANEL-clone"></a>}}
\if{latex}{\out{\hypertarget{method-PosteriorBVARPANEL-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{specify_posterior_bvarPANEL$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
