% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.R
\name{summary.ForecastsPANEL}
\alias{summary.ForecastsPANEL}
\title{Provides posterior summary of country-specific Forecasts}
\usage{
\method{summary}{ForecastsPANEL}(object, which_c, ...)
}
\arguments{
\item{object}{an object of class \code{ForecastsPANEL} obtained using the
\code{forecast()} function containing draws the predictive density.}

\item{which_c}{a positive integer or a character string specifying the country 
for which the forecast should be plotted.}

\item{...}{additional arguments affecting the summary produced.}
}
\value{
A list reporting the posterior mean, standard deviations, as well as 
5 and 95 percentiles of the forecasts for each of the variables and forecast 
horizons.
}
\description{
Provides posterior summary of the forecasts including their 
mean, standard deviations, as well as 5 and 95 percentiles.
}
\examples{
data(ilo_cubic_panel)                                   # load the data
data(ilo_exogenous_variables)                           # load the exogenous variables
data(ilo_exogenous_forecasts)                           # load the exogenous forecast
set.seed(123)

# specify the model
specification = specify_bvarPANEL$new(ilo_cubic_panel, exogenous = ilo_exogenous_variables)
burn_in       = estimate(specification, 10)             # run the burn-in
posterior     = estimate(burn_in, 10)                   # estimate the model

# forecast 6 years ahead
predictive    = forecast(posterior, 6, exogenous_forecast = ilo_exogenous_forecasts)
summary(predictive, which_c = "POL")

# workflow with the pipe |>
############################################################
set.seed(123)
ilo_cubic_panel |>
  specify_bvarPANEL$new() |>
  estimate(S = 10) |> 
  estimate(S = 20) |> 
  forecast(horizon = 2) |> 
  summary(which_c = "POL")

# conditional forecasting 6 years ahead conditioning on 
#  provided future values for the Gross Domestic Product 
#  growth rate
############################################################
data(ilo_conditional_forecasts)                        # load the conditional forecasts of dgdp
specification = specify_bvarPANEL$new(ilo_cubic_panel)    # specify the model
burn_in       = estimate(specification, 10)               # run the burn-in
posterior     = estimate(burn_in, 10)                     # estimate the model
# forecast 6 years ahead
predictive    = forecast(posterior, 6, conditional_forecast = ilo_conditional_forecasts)
summary(predictive, which_c = "POL")

# workflow with the pipe |>
############################################################
set.seed(123)
ilo_cubic_panel |>
  specify_bvarPANEL$new() |>
  estimate(S = 10) |> 
  estimate(S = 20) |> 
  forecast(
    horizon = 6, 
    conditional_forecast = ilo_conditional_forecasts
  ) |> 
  summary(which_c = "POL")

}
\seealso{
\code{\link{forecast.PosteriorBVARPANEL}}, \code{\link{plot}}
}
\author{
Tomasz Woźniak \email{wozniak.tom@pm.me}
}
