% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.R
\name{summary.PosteriorBVARPANEL}
\alias{summary.PosteriorBVARPANEL}
\title{Provides posterior estimation summary for Bayesian Hierarchical Panel 
Vector Autoregressions}
\usage{
\method{summary}{PosteriorBVARPANEL}(object, ...)
}
\arguments{
\item{object}{an object of class \code{PosteriorBVARPANEL} obtained using the
\code{estimate()} function applied to  
Vector Autoregressions containing draws from the  posterior distribution of 
the parameters.}

\item{...}{additional arguments affecting the summary produced.}
}
\value{
A list reporting the posterior mean, standard deviations, as well as 5 and 95 
percentiles of the country-specific and global parameters.
}
\description{
Provides posterior mean, standard deviations, as well as 5 and 95 
percentiles of the parameters for all \code{C} countries.
}
\examples{
# upload data
data(ilo_cubic_panel)                                   # load the data
data(ilo_exogenous_variables)                           # load the exogenous variables

set.seed(123)

# specify the model
specification = specify_bvarPANEL$new(ilo_cubic_panel, exogenous = ilo_exogenous_variables)
burn_in       = estimate(specification, 10)             # run the burn-in
posterior     = estimate(burn_in, 10)                   # estimate the model
summary(posterior)

# workflow with the pipe |>
############################################################
set.seed(123)
ilo_cubic_panel |>
  specify_bvarPANEL$new(exogenous = ilo_exogenous_variables) |>
  estimate(S = 10) |> 
  estimate(S = 10) |> 
  summary()

}
\seealso{
\code{\link{estimate.BVARPANEL}}, \code{\link{specify_bvarPANEL}}
}
\author{
Tomasz Woźniak \email{wozniak.tom@pm.me}
}
