% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bvarPANELs-package.R
\docType{package}
\name{bvarPANELs-package}
\alias{bvarPANELs-package}
\alias{bvarPANELs}
\title{Forecasting with Bayesian Hierarchical Panel Vector Autoregressions}
\description{
Forecasting a multi-country time series panel data using 
Bayesian Vector Autoregressions with a three-level country-global 
hierarchical prior structure. Copyright: 2024 International Labour Organization.
}
\note{
This package is currently in active development.
}
\examples{
# Basic estimation and forecasting example
############################################################
data(ilo_cubic_panel)                                   # load the data
set.seed(123)
specification = specify_bvarPANEL$new(ilo_cubic_panel)  # specify the model
burn_in       = estimate(specification, 10)             # run the burn-in
posterior     = estimate(burn_in, 10)                   # estimate the model
predictive    = forecast(posterior, 2)                  # forecast the future       

# workflow with the pipe |>
set.seed(123)
ilo_cubic_panel |>
  specify_bvarPANEL$new() |>
  estimate(S = 20) |> 
  estimate(S = 20) |> 
  forecast(horizon = 2) -> predictive
  
# Full estimation and forecasting example with 
#   exogenous variables and conditional forecasts
############################################################
data(ilo_cubic_panel)                                   # load the data
data(ilo_exogenous_variables)                           # load the exogenous variables
data(ilo_exogenous_forecasts)                           # load the exogenous forecasts
data(ilo_conditional_forecasts)                         # load the conditional forecasts
set.seed(123)
specification = specify_bvarPANEL$new(
                  ilo_cubic_panel,
                  exogenous = ilo_exogenous_variables
                )
burn_in       = estimate(specification, 10)             # run the burn-in
posterior     = estimate(burn_in, 10)                   # estimate the model
predictive    = forecast(
                  posterior, 
                  horizon = 6,
                  exogenous_forecast = ilo_exogenous_forecasts,
                  conditional_forecast = ilo_conditional_forecasts
                )
}
\author{
Tomasz Woźniak \email{wozniak.tom@pm.me}
}
\keyword{#'}
\keyword{@examples}
\keyword{models}
\keyword{package}
\keyword{ts}
