% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bvarPANELs-package.R
\docType{package}
\name{bvarPANELs-package}
\alias{bvarPANELs-package}
\alias{bvarPANELs}
\title{Forecasting with Bayesian Hierarchical Panel Vector Autoregressions}
\description{
Forecasting a multi-country time series panel data using 
Bayesian Vector Autoregressions with a three-level country-global 
hierarchical prior structure. Copyright: 2024 International Labour Organization.
}
\note{
This package is currently in active development.
}
\author{
Tomasz Woźniak \email{wozniak.tom@pm.me}
}
\keyword{#}
\keyword{#'}
\keyword{10)}
\keyword{=}
\keyword{@examples}
\keyword{burn-in}
\keyword{burn_in}
\keyword{data}
\keyword{data(ilo_cubic_panel)}
\keyword{estimate}
\keyword{estimate(burn_in,}
\keyword{estimate(specification,}
\keyword{load}
\keyword{model}
\keyword{models}
\keyword{package}
\keyword{posterior}
\keyword{run}
\keyword{set.seed(123)}
\keyword{specification}
\keyword{specify}
\keyword{specify_bvarPANEL$new(ilo_cubic_panel)}
\keyword{the}
\keyword{ts}
