% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forecast.R
\name{forecast.PosteriorBVARPANEL}
\alias{forecast.PosteriorBVARPANEL}
\title{Forecasting using Hierarchical Panel Vector Autoregressions}
\usage{
\method{forecast}{PosteriorBVARPANEL}(
  posterior,
  horizon = 1,
  exogenous_forecast = NULL,
  conditional_forecast = NULL
)
}
\arguments{
\item{posterior}{posterior estimation outcome - an object of class 
PosteriorBVARPANEL obtained by running the \code{estimate} function.}

\item{horizon}{a positive integer, specifying the forecasting horizon.}

\item{exogenous_forecast}{not used here ATM; included for compatibility with 
generic \code{forecast}.}

\item{conditional_forecast}{a list of length \code{C} containing 
\code{horizon x N} matrices with forecasted values for selected variables. 
These matrices should only contain \code{numeric} or \code{NA} values. The 
entries with \code{NA} values correspond to the values that are forecasted 
conditionally on the realisations provided as \code{numeric} values.}
}
\value{
A list of class \code{ForecastsPANEL} with \code{C} elements containing 
the draws from the country-specific predictive density and data in a form of 
object class \code{Forecasts} that includes:

\describe{
 \item{forecasts}{an \code{horizonxNxS} array with the draws from the 
 country-specific predictive density}
 \item{Y}{a \code{T_cxN} matrix with the country-specific data}
}
}
\description{
Samples from the joint predictive density of the dependent 
variables for all countries at forecast horizons 
from 1 to \code{horizon} specified as an argument of the function.
}
\examples{
data(ilo_cubic_panel)                                   # load the data
data(ilo_exogenous_variables)                           # load the exogenous variables
data(ilo_exogenous_forecasts)                           # load the exogenous forecast
set.seed(123)

# specify the model
specification = specify_bvarPANEL$new(ilo_cubic_panel, exogenous = ilo_exogenous_variables)
burn_in       = estimate(specification, 10)             # run the burn-in
posterior     = estimate(burn_in, 10)                   # estimate the model

# forecast 6 years ahead
predictive    = forecast(posterior, 6, exogenous_forecast = ilo_exogenous_forecasts)

# workflow with the pipe |>
############################################################
set.seed(123)
ilo_cubic_panel |>
  specify_bvarPANEL$new() |>
  estimate(S = 10) |> 
  estimate(S = 20) |> 
  forecast(horizon = 2) -> predictive

# conditional forecasting 6 years ahead conditioning on 
#  provided future values for the Gross Domestic Product 
#  growth rate
############################################################
data(ilo_conditional_forecasts)                        # load the conditional forecasts of dgdp
specification = specify_bvarPANEL$new(ilo_cubic_panel)    # specify the model
burn_in       = estimate(specification, 10)               # run the burn-in
posterior     = estimate(burn_in, 10)                     # estimate the model
# forecast 6 years ahead
predictive    = forecast(posterior, 6, conditional_forecast = ilo_conditional_forecasts)

# workflow with the pipe |>
############################################################
set.seed(123)
ilo_cubic_panel |>
  specify_bvarPANEL$new() |>
  estimate(S = 10) |> 
  estimate(S = 20) |> 
  forecast(
    horizon = 6, 
    conditional_forecast = ilo_conditional_forecasts
  ) -> predictive

}
\seealso{
\code{\link{estimate.PosteriorBVARPANEL}}
}
\author{
Tomasz Woźniak \email{wozniak.tom@pm.me}
}
