% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/specify_bvarpanel.R
\name{specify_bvarPANEL}
\alias{specify_bvarPANEL}
\title{R6 Class representing the specification of the BVARPANEL model}
\description{
The class BVARPANEL presents complete specification for the Bayesian Panel
Vector Autoregression.
}
\examples{
\dontrun{
data(us_fiscal_lsuw)
spec = specify_bvarPANEL$new(
   data = us_fiscal_lsuw,
   p = 4
)
}


## ------------------------------------------------
## Method `specify_bvarPANEL$get_data_matrices`
## ------------------------------------------------

\dontrun{ 
data(us_fiscal_lsuw)
spec = specify_bsvar$new(
   data = us_fiscal_lsuw,
   p = 4
)
spec$get_data_matrices()
}

## ------------------------------------------------
## Method `specify_bvarPANEL$get_prior`
## ------------------------------------------------

\dontrun{
data(us_fiscal_lsuw)
spec = specify_bsvar$new(
   data = us_fiscal_lsuw,
   p = 4
)
spec$get_prior()
}

## ------------------------------------------------
## Method `specify_bvarPANEL$get_starting_values`
## ------------------------------------------------

\dontrun{
data(us_fiscal_lsuw)
spec = specify_bsvar$new(
   data = us_fiscal_lsuw,
   p = 4
)
spec$get_starting_values()
}
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{p}}{a non-negative integer specifying the autoregressive lag order of the model.}

\item{\code{prior}}{an object PriorBSVAR with the prior specification.}

\item{\code{data_matrices}}{an object DataMatricesBVARPANEL with the data matrices.}

\item{\code{starting_values}}{an object StartingValuesBVARPANEL with the starting values.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-BVARPANEL-new}{\code{specify_bvarPANEL$new()}}
\item \href{#method-BVARPANEL-get_data_matrices}{\code{specify_bvarPANEL$get_data_matrices()}}
\item \href{#method-BVARPANEL-get_prior}{\code{specify_bvarPANEL$get_prior()}}
\item \href{#method-BVARPANEL-get_starting_values}{\code{specify_bvarPANEL$get_starting_values()}}
\item \href{#method-BVARPANEL-clone}{\code{specify_bvarPANEL$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-BVARPANEL-new"></a>}}
\if{latex}{\out{\hypertarget{method-BVARPANEL-new}{}}}
\subsection{Method \code{new()}}{
Create a new specification of the Bayesian Panel VAR model BVARPANEL.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{specify_bvarPANEL$new(data, p = 1L)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{data}}{a list with \code{C} elements of \code{(T_c+p)xN} matrices 
with time series data.}

\item{\code{p}}{a positive integer providing model's autoregressive lag order.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A new complete specification for the Bayesian Panel VAR model BVARPANEL.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-BVARPANEL-get_data_matrices"></a>}}
\if{latex}{\out{\hypertarget{method-BVARPANEL-get_data_matrices}{}}}
\subsection{Method \code{get_data_matrices()}}{
Returns the data matrices as the DataMatricesBVARPANEL object.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{specify_bvarPANEL$get_data_matrices()}\if{html}{\out{</div>}}
}

\subsection{Examples}{
\if{html}{\out{<div class="r example copy">}}
\preformatted{\dontrun{ 
data(us_fiscal_lsuw)
spec = specify_bsvar$new(
   data = us_fiscal_lsuw,
   p = 4
)
spec$get_data_matrices()
}
}
\if{html}{\out{</div>}}

}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-BVARPANEL-get_prior"></a>}}
\if{latex}{\out{\hypertarget{method-BVARPANEL-get_prior}{}}}
\subsection{Method \code{get_prior()}}{
Returns the prior specification as the PriorBVARPANEL object.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{specify_bvarPANEL$get_prior()}\if{html}{\out{</div>}}
}

\subsection{Examples}{
\if{html}{\out{<div class="r example copy">}}
\preformatted{\dontrun{
data(us_fiscal_lsuw)
spec = specify_bsvar$new(
   data = us_fiscal_lsuw,
   p = 4
)
spec$get_prior()
}
}
\if{html}{\out{</div>}}

}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-BVARPANEL-get_starting_values"></a>}}
\if{latex}{\out{\hypertarget{method-BVARPANEL-get_starting_values}{}}}
\subsection{Method \code{get_starting_values()}}{
Returns the starting values as the StartingValuesBVARPANEL object.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{specify_bvarPANEL$get_starting_values()}\if{html}{\out{</div>}}
}

\subsection{Examples}{
\if{html}{\out{<div class="r example copy">}}
\preformatted{\dontrun{
data(us_fiscal_lsuw)
spec = specify_bsvar$new(
   data = us_fiscal_lsuw,
   p = 4
)
spec$get_starting_values()
}
}
\if{html}{\out{</div>}}

}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-BVARPANEL-clone"></a>}}
\if{latex}{\out{\hypertarget{method-BVARPANEL-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{specify_bvarPANEL$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
